Rolled back matplotlib.use("Agg") change which broke catreport - seems to be working without it now anyway.
Added Age units for generating animations of age output layers; graph will show the area-weighted average age of the whole landscape, and the map will show exact pixel age.

Made LoopingIterator an actual iterator and fixed infinite loop bug in Palette.

+++ catplotlib/spatial/layer.py
@@ -222,7 +222,7 @@ class Layer:
 
         raster = gdal.Open(self._path)
         band = raster.GetRasterBand(1)
-        raster_data = band.ReadAsArray()
+        raster_data = band.ReadAsArray().astype("float")
         nodata_value = self.nodata_value
 
         # First need to convert pixel values to absolute (as opposed to per hectare), in the correct
